CDSI_construction_scripts/requirements.txt
sahi==0.11.14
fire==0.5.0
scipy==1.10.1
tqdm==4.65.0
pillow
dotadevkit==1.3.0
pyodi==0.0.7
